finalized cashing for comment and blog

+++ backend/AAiT-backend-group-8/Infrastructure/mongodb/comment.go
@@ -4,7 +4,6 @@ import (
 	domain "AAiT-backend-group-8/Domain"
 	"context"
 	"errors"
-
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
@@ -50,16 +49,26 @@ func (repo *CommentRepository) GetComments(blogID primitive.ObjectID) ([]domain.
 	return comments, nil
 }
 
-func (repo *CommentRepository) DeleteComment(commentID primitive.ObjectID) error {
+func (repo *CommentRepository) DeleteComment(commentID primitive.ObjectID) (string, error) {
 	filter := bson.D{{Key: "_id", Value: commentID}}
+	findResult := repo.collection.FindOne(repo.ctx, filter)
+	var comment domain.Comment
+
+	if findResult.Decode(&comment) != nil {
+		//fmt.Println("error deleting comment")
+		return "", errors.New("error deleting comment")
+	}
+
 	result, err := repo.collection.DeleteOne(repo.ctx, filter)
+
 	if err != nil {
-		return errors.New("error deleting comment")
+		return "", errors.New("error deleting comment")
 	}
 	if result.DeletedCount == 0 {
-		return errors.New("comment not found")
+		return "", errors.New("comment not found")
 	}
-	return nil
+
+	return comment.BlogID.Hex(), nil
 }
 
 func (repo *CommentRepository) DeleteCommentsOfBlog(blogID primitive.ObjectID) error {
@@ -74,19 +83,31 @@ func (repo *CommentRepository) DeleteCommentsOfBlog(blogID primitive.ObjectID) e
 	return nil
 }
 
-func (repo *CommentRepository) UpdateComment(comment *domain.Comment) error {
+func (repo *CommentRepository) UpdateComment(comment *domain.Comment) (string, error) {
 	filter := bson.D{{Key: "_id", Value: comment.Id}}
+
 	update := bson.D{{Key: "$set", Value: bson.D{
 		{Key: "body", Value: comment.Body},
 	}}}
+
+	var commnt domain.Comment
+
+	findOneResult := repo.collection.FindOne(repo.ctx, filter)
+
+	if findOneResult.Decode(&commnt) != nil {
+		return "", errors.New("error updating comment")
+	}
+
 	updated, err := repo.collection.UpdateOne(repo.ctx, filter, update)
+
 	if err != nil {
-		return errors.New("error updating comment")
+		return "", errors.New("error updating comment")
 	}
+
 	if updated.ModifiedCount == 0 {
-		return errors.New("comment not found")
+		return "", errors.New("comment not found")
 	}
-	return nil
+	return commnt.BlogID.Hex(), nil
 }
 
 func (repo *CommentRepository) GetCommentByID(commentID primitive.ObjectID) (*domain.Comment, error) {
